+++ frontend/src/features/main.ts
@@ -1,5 +1,8 @@
 import "../styles/style.css"
 
+
 console.log("Hello from TypeScript!");
 
-import "./loginRedirect"; // Import du script de redirection
+//Routing
+import { setupRouting } from "./router";
+setupRouting();
